GC2D/CardLoad and CardSave: match the score screen, drawMessage and waitForAnyKey (#180)
* GC2D/CardLoad and CardSave: match the score screen, drawMessage and waitForAnyKey

* GC2D/CardLoad: drop the frame note in changeScene

* GC2D/CardSave: keep the getCurMessageID getter

+++ src/GC2D/CardSave.cpp
@@ -955,7 +955,7 @@ s8 TCardSave::drawMessageBM(TEProgress param_1)
 
 s8 TCardSave::waitForAnyKey(TEProgress param_1)
 {
-	s32 result = -1;
+	s8 result = -1;
 
 	switch (unk10) {
 	case 0:
